requirements.txt
flask
pandas
requests
sentence-transformers
faiss-cpu
numpy
python-dotenv